--- Components/sliderIntro.js
@@ -1,8 +1,0 @@
-import React, {useState} from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import AppIntroSlider from 'react-native-app-intro-slider';
-
-
-export default function Slider() {
-
-}
