+++ README.md
@@ -1,2 +1,21 @@
 # CONPASS
-campus map mobile app project for mini-capstone course at Concordia University 
+Campus map mobile app project for mini-capstone course at Concordia University 
+
+## Setting up
+
+1. install expo cli on your machine <br/>
+`npm install -g expo-cli`
+2. install dependencies <br/>
+`yarn install`
+3. start app (see package.json scripts) <br/>
+`expo start`
+
+
+## Good to know
+1. added package.json file within src/components directory in order to absolute reference. [ref](https://medium.com/better-programming/using-absolute-paths-in-react-native-3be369244fb1)
+
+2. we are going to use Yarn for dependency installation
+
+
+## Useful links
+[styling in React Native](https://medium.com/mindorks/everything-to-know-about-styling-in-react-native-7e30aed53ad)
