+++ package.json
@@ -1,4 +1,5 @@
 {
+  
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
