+++ src/components/home/styles.js
@@ -1,0 +1,17 @@
+import { StyleSheet} from 'react-native';  
+
+const styles = StyleSheet.create({
+  container: {  
+    flex: 1,  
+    justifyContent: 'center',  
+    alignItems: 'center',  
+    backgroundColor: '#F5FCFF',  
+  },  
+  welcome: {  
+    fontSize: 20,  
+    textAlign: 'center',  
+    margin: 10,  
+  },  
+  });
+
+export default styles;
