+++ src/components/package.json
@@ -1,0 +1,3 @@
+{
+    "name": "components"
+}
